requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.0
reportlab>=4.0.0
google-generativeai>=0.3.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
